+++ govtool/backend/src/VVA/Config.hs
@@ -29,6 +29,7 @@ module VVA.Config
     , vvaConfigToText
     , getMetadataValidationHost
     , getMetadataValidationPort
+    , getWebsocketLifetimeSeconds
     ) where
 
 import           Conferer
@@ -101,6 +102,8 @@ data VVAConfigInternal
       , vVAConfigInternalMetadataValidationMaxConcurrentRequests :: Int
         -- | Redis config
       , vVAConfigInternalRedisConfig :: RedisInternalConfig
+        -- | WebSocket lifetime in seconds
+      , vVAConfigInternalWebsocketLifetimeSeconds :: Int
       }
   deriving (FromConfig, Generic, Show)
 
@@ -116,7 +119,8 @@ instance DefaultConfig VVAConfigInternal where
         vVAConfigInternalMetadataValidationHost = "localhost",
         vVAConfigInternalMetadataValidationPort = 3001,
         vVAConfigInternalMetadataValidationMaxConcurrentRequests = 10,
-        vVAConfigInternalRedisConfig = RedisInternalConfig "localhost" 6379 Nothing
+        vVAConfigInternalRedisConfig = RedisInternalConfig "localhost" 6379 Nothing,
+        vVAConfigInternalWebsocketLifetimeSeconds = 60 * 1
       }
 
 data RedisConfig
@@ -150,6 +154,8 @@ data VVAConfig
       , metadataValidationMaxConcurrentRequests :: Int
         -- | Redis config
       , redisConfig :: RedisConfig
+        -- | WebSocket lifetime in seconds
+      , websocketLifetimeSeconds :: Int
       }
   deriving (Generic, Show, ToJSON)
 
@@ -199,6 +205,7 @@ convertConfig VVAConfigInternal {..} =
           redisPort = redisInternalConfigPort $ vVAConfigInternalRedisConfig,
           redisPassword = redisInternalConfigPassword $ vVAConfigInternalRedisConfig
         }
+      websocketLifetimeSeconds = vVAConfigInternalWebsocketLifetimeSeconds
     }
 
 -- | Load configuration from a file specified on the command line.  Load from
@@ -265,3 +272,9 @@ getMetadataValidationPort ::
   (Has VVAConfig r, MonadReader r m) =>
   m Int
 getMetadataValidationPort = asks (metadataValidationPort . getter)
+
+-- | Access websocket lifetime in seconds
+getWebsocketLifetimeSeconds ::
+  (Has VVAConfig r, MonadReader r m) =>
+  m Int
+getWebsocketLifetimeSeconds = asks (websocketLifetimeSeconds . getter)